refactor: Move chart related components to single folder
