+++ components/charts/Bar.tsx
@@ -2,7 +2,7 @@ import	React		from	'react';
 import {Bar, BarChart, ResponsiveContainer, Tooltip, XAxis, YAxis} from 'recharts';
 import {formatXAxis, formatYAxis} from 'utils/b2b/Chart';
 
-import CustomTooltip from '../CustomTooltip';
+import CustomTooltip from './CustomTooltip';
 
 import type {ReactElement} from 'react';
 import type {TChartProps} from 'types/chart';
