+++ components/charts/ChartLegend.tsx
