+++ components/charts/Composed.tsx
@@ -2,7 +2,7 @@ import	React, {useState}		from	'react';
 import {Bar, Cell, ComposedChart, ResponsiveContainer, Tooltip, XAxis, YAxis} from 'recharts';
 import {formatXAxis, formatYAxis} from 'utils/b2b/Chart';
 
-import CustomTooltip from '../CustomTooltip';
+import CustomTooltip from './CustomTooltip';
 
 import type {ReactElement} from 'react';
 import type {TChartProps} from 'types/chart';
